focaltech: Reorder MT_PRESSURE and MT_TOOL_FINGER
Report the MT_PRESSURE to be zero before marking the slot as unused.
Android's InputReader marks the slot as no longer in use when the
driver sets MT_TRACKING_ID to -1. Further MT events will re-mark
the slot as in-use, so we should ensure that we complete all
coordinate/pressure events before the final MT_TRACKING_ID event.

Change-Id: Ia4529b2815e3eda836fa224c859732698f37cefa
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>